+++ src/OpenGL/SSBO.h
@@ -13,7 +13,7 @@ public:
 		GLCall(glDeleteBuffers(1, &m_id));
 	}
 
-	void uploadData(int nbOfT, T* data) {
+	void uploadData(size_t nbOfT, T* data) {
 		GLCall(glBindBuffer(GL_SHADER_STORAGE_BUFFER, m_id));
 		GLCall(glBufferData(GL_SHADER_STORAGE_BUFFER, nbOfT * sizeof(T), data, m_hint));
 	}
